+++ docs/demo/select/section-async-multiple.jsx
@@ -2,39 +2,74 @@ import React from 'react'
 import DocViewer from '../../../libs/doc-viewer'
 import Select from '../../../components/select'
 const prefix = 'select-async-multiple'
-const code = `import React from 'react'
-import { Select } from '@hi-ui/hiui'\n
-class Demo extends React.Component {
-  render () {
-    return (
-      <Select
-        type='multiple'
-        autoload
-        style={{ width: 300 }}
-        defaultValue={1}
-        dataSource={{
-          type: 'get',
-          key: 'text',
-          keyword: 'xiaomi',
-          url: 'https://easy-mock.com/mock/5c1b42e3fe5907404e6540e9/hiui/select/options',
-          transformResponse: (res) => {
-            console.log('----', res)
-            return res.data
-          },
-          error: err => console.log('error:', err)
-        }}
-        onChange={(item) => {
-          console.log('异步多选结果', item)
-        }}
-      />
-    )
-  }
-}`
+const rightOptions = ['使用DataSource', '使用onSearch']
+const code = [{
+  code: `import React from 'react'
+    import { Select } from '@hi-ui/hiui'\n
+    class Demo extends React.Component {
+      render () {
+        return (
+          <Select
+          type='multiple'
+          style={{width: '300px'}}
+          placeholder='请选择...'
+          dataSource={{
+            type: 'GET',
+            url: 'http://yapi.demo.qunar.com/mock/26534/hiui/select',
+            key: 'id',
+            transformResponse: (res) => {
+              return res.list
+            }
+          }}
+          onChange={(item) => {
+            console.log('多选结果', item)
+          }}
+          />
+        )
+      }
+    }`,
+  opt: ['使用DataSource']
+}, {
+  code: `import React from 'react'
+    import { Select } from '@hi-ui/hiui'\n
+    class Demo extends React.Component {
+      constructor () {
+        super()
+        this.state = {
+          multipleList: []
+        }
+      }
+      render () {
+        return (
+          <Select
+          type='multiple'
+          style={{width: '300px'}}
+          data={this.state.multipleList}
+          placeholder='请选择...'
+          onSearch={(value) => {
+            return fetch('http://yapi.demo.qunar.com/mock/26534/hiui/select?id='+value+'')
+            .then(res => res.json())
+            .then(res => {
+              this.setState({
+                multipleList: res.list
+              })
+            })
+          }}
+          onChange={(item) => {
+            console.log('多选结果', item)
+          }}
+          />
+        )
+      }
+    }`,
+  opt: ['使用onSearch']
+}]
 const DemoSingleMultiple = () => (
   <DocViewer
     code={code}
     scope={{ Select }}
     prefix={prefix}
+    rightOptions={rightOptions}
   />
 )
 export default DemoSingleMultiple