+++ docs/demo/tree/section-async.jsx
@@ -43,7 +43,7 @@ class Demo extends React.Component {
             headers:{},
             data:{},
             params:{},
-            url:'https://easy-mock.com/mock/5c1b42e3fe5907404e6540e9/hiui/select/options',
+            url:'http://yapi.demo.qunar.com/mock/26534/hiui/async-tree',
             transformResponse:(res)=>{return res.data}
           }}
           defaultExpandAll