+++ docs/demo/upload/section-ban.jsx
@@ -10,7 +10,7 @@ class Demo extends React.Component {
     return (
       <Upload
         type="normal"
-        uploadAction= "https://easy-mock.com/mock/5c1b42e3fe5907404e6540e9/hiui/upload"
+        uploadAction= "http://www.mocky.io/v2/5dc3b4413000007600347501"
         buttonText="上传文件"
         params={{id:'uid',channel:'youpin'}}
         disabled={true}