+++ docs/demo/upload/section-default.jsx
@@ -15,7 +15,7 @@ class Demo extends React.Component {
       <div>
         <Upload
           type="normal"
-          uploadAction= "https://easy-mock.com/mock/5c1b42e3fe5907404e6540e9/hiui/upload"
+          uploadAction= "http://www.mocky.io/v2/5dc3b4413000007600347501"
           buttonText="上传文件"
           params={param}
           name={'files[]'}